Created pre_fuzzy_cleansed.sql script
This commit provides a dbt model that performs data cleansing operations on site and parent account names. Removed non-alphanumeric characters and applied a function to provide uniform uppercase formatting. The cleaned data is included in the final output for each unique Account_UUID, along with the country and source details.